[Feature] Full announcement API revamp

+++ src/controllers/announcement.controller.js
@@ -7,12 +7,15 @@ const createAnnouncement = async (req, res) => {
     name,
     number,
     email,
+    age,
+    about,
     type,
     genreIds,
     state,
     city,
     description,
     instrumentIds,
+    socialLinks,
   } = req.body;
 
   if (
@@ -20,6 +23,8 @@ const createAnnouncement = async (req, res) => {
     !name ||
     !number ||
     !email ||
+    !age ||
+    !about ||
     !type ||
     !genreIds ||
     !state ||
@@ -28,8 +33,7 @@ const createAnnouncement = async (req, res) => {
     !instrumentIds
   ) {
     return res.status(400).send({
-      message:
-        "One or more required fields are missing: Title, Name, Number, Email, Type, GenreIds, State, City, Description, InstrumentIds",
+      message: "One or more required fields are missing",
     });
   }
 
@@ -44,12 +48,25 @@ const createAnnouncement = async (req, res) => {
         state,
         city,
         description,
+        age,
+        about,
         genres: {
           connect: genreIds.map((genreId) => ({ id: genreId })),
         },
         instruments: {
           connect: instrumentIds.map((instrumentId) => ({ id: instrumentId })),
         },
+        socialLinks: {
+          create: socialLinks.map(({ socialMediaId, url }) => ({
+            url,
+            socialMedia: { connect: { id: socialMediaId } },
+          })),
+        },
+      },
+      include: {
+        genres: true,
+        instruments: true,
+        socialLinks: { include: { socialMedia: true } },
       },
     });
 
@@ -58,29 +75,27 @@ const createAnnouncement = async (req, res) => {
       data: announcement,
     });
   } catch (err) {
-    if (err instanceof PrismaClientKnownRequestError) {
-      if (err.code === "P2002") {
-        return res.status(409).json({
-          status: "error",
-          message: "Duplicate entry found for unique field.",
-        });
-      }
-    }
-
     console.error(err);
     return res.status(500).json({
       status: "error",
-      message:
-        "An unexpected error occurred while creating the announcement. Please try again later.",
+      message: "An error occurred while creating the announcement.",
     });
   }
 };
 
 const deleteAnnouncement = async (req, res) => {
-  const id = req.params.id;
+  const id = parseInt(req.params.id); // Convert the ID to an integer
 
   try {
-    const announcement = await prisma.announcement.delete({
+    const existingAnnouncement = await prisma.announcement.findUnique({
+      where: { id },
+    });
+
+    if (!existingAnnouncement) {
+      return res.status(404).json({ message: "Announcement not found" });
+    }
+
+    await prisma.announcement.delete({
       where: { id: parseInt(id) },
     });
     res.status(200).send({ message: "Announcement deleted successfully." });
@@ -93,18 +108,21 @@ const deleteAnnouncement = async (req, res) => {
 };
 
 const updateAnnouncement = async (req, res) => {
-  const id = req.params.id;
+  const id = parseInt(req.params.id);
   const {
     title,
     name,
     number,
     email,
+    age,
+    about,
     type,
     genreIds,
     state,
     city,
     description,
     instrumentIds,
+    socialLinks,
   } = req.body;
 
   if (
@@ -112,6 +130,8 @@ const updateAnnouncement = async (req, res) => {
     !name ||
     !number ||
     !email ||
+    !age ||
+    !about ||
     !type ||
     !genreIds ||
     !state ||
@@ -126,35 +146,74 @@ const updateAnnouncement = async (req, res) => {
   }
 
   try {
+    const existingAnnouncement = await prisma.announcement.findUnique({
+      where: { id },
+    });
+
+    if (!existingAnnouncement) {
+      return res.status(404).json({ message: "Announcement not found" });
+    }
+
     const updatedAnnouncement = await prisma.announcement.update({
-      where: { id: parseInt(id) },
+      where: { id },
       data: {
         title,
-        genre,
+        name,
+        number,
+        email,
+        type,
+        age,
+        about,
         state,
         city,
         description,
-        owner_type: ownerType,
-        owner_id: ownerId,
-        latitude,
-        longitude,
+        genres: {
+          set: genreIds.map((genreId) => ({ id: genreId })),
+        },
+        instruments: {
+          set: instrumentIds.map((instrumentId) => ({ id: instrumentId })),
+        },
+        socialLinks: {
+          deleteMany: {},
+          create: socialLinks.map(({ socialMediaId, url }) => ({
+            url,
+            socialMedia: { connect: { id: socialMediaId } },
+          })),
+        },
+      },
+      include: {
+        genres: true,
+        instruments: true,
+        socialLinks: { include: { socialMedia: true } },
       },
     });
-    res.send({
-      message: "Announcement updated successfully.",
+
+    return res.status(200).json({
+      status: "success",
       data: updatedAnnouncement,
     });
   } catch (err) {
-    res.status(500).send({
-      message:
-        err.message || "We had some errors while updating this announcement.",
+    console.error(err);
+    return res.status(500).json({
+      status: "error",
+      message: "An error occurred while updating the announcement.",
     });
   }
 };
 
 const getAnnouncements = async (req, res) => {
   try {
-    const announcementsList = await prisma.announcement.findMany();
+    const announcementsList = await prisma.announcement.findMany({
+      include: {
+        genres: true,
+        instruments: true,
+      },
+    });
+    announcementsList.forEach((announcement) => {
+      delete announcement.about;
+      delete announcement.email;
+      delete announcement.number;
+    });
     res.send(announcementsList);
   } catch (err) {
     res.status(500).send({
@@ -164,9 +223,55 @@ const getAnnouncements = async (req, res) => {
   }
 };
 
+const getAnnouncementById = async (req, res) => {
+  const id = parseInt(req.params.id);
+
+  try {
+    const announcement = await prisma.announcement.findUnique({
+      where: { id },
+      include: {
+        socialLinks: {
+          include: {
+            socialMedia: {
+              select: {
+                id: true,
+                name: true,
+              },
+            },
+          },
+        },
+        genres: true,
+        instruments: true,
+      },
+    });
+
+    if (!announcement) {
+      return res.status(404).json({ message: "Announcement not found" });
+    }
+
+    const socialLinks = announcement.socialLinks.map((socialLink) => ({
+      id: socialLink.id,
+      url: socialLink.url,
+      socialMediaId: socialLink.socialMedia.id,
+      socialMediaName: socialLink.socialMedia.name,
+    }));
+
+    res.status(200).json({
+      ...announcement,
+      socialLinks,
+    });
+  } catch (err) {
+    res.status(500).send({
+      message:
+        err.message || "Some error occurred while retrieving the announcement.",
+    });
+  }
+};
+
 export const announcementController = {
   createAnnouncement,
   deleteAnnouncement,
   updateAnnouncement,
   getAnnouncements,
+  getAnnouncementById,
 };
